fastapi-backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
pandas
scikit-learn
